requirements.txt
Django==1.8.15
django-cors-headers==1.2.2
django-filter==0.15.2
djangorestframework==3.4.7
Markdown==2.6.7
oauthlib==2.0.0
Pillow==3.3.1
requests==2.11.1
requests-oauthlib==0.7.0
